requirements.txt
matplotlib==3.1.2
numpy==1.17.4
pandas==1.2.3
Pillow==9.1.0
scikit_learn==1.0.2
tensorflow==2.4.1
